Restructure and rename files
