+++ public/chrome_manifest.json
@@ -9,12 +9,12 @@
     },
     "content_scripts": [
         {
-            "js": ["browser-polyfill.js", "contentscript.js"],
+            "js": ["browser-polyfill.js", "content.js"],
             "matches": ["https://www.desmos.com/calculator/*"],
             "run_at": "document_idle"
         },
         {
-            "js": ["browser-polyfill.js", "preload_content.js"],
+            "js": ["browser-polyfill.js", "preloadContent.js"],
             "matches": ["https://www.desmos.com/calculator*"],
             "run_at": "document_start"
         }
@@ -22,7 +22,7 @@
     "host_permissions": ["*://www.desmos.com/calculator/*"],
     "web_accessible_resources": [
         {
-            "resources": ["browser-polyfill.js", "inject.js", "preload_inject.js", "run_calculator.js", "empty.js"],
+            "resources": ["browser-polyfill.js", "script.js", "preloadScript.js", "run_calculator.js", "empty.js"],
             "matches": ["<all_urls>"]
         }
     ],
