+++ src/content.ts
@@ -1,7 +1,7 @@
 async function updateAutoCommands() {
     const commands = await browser.storage.local.get('autoCommands');
     const script = document.createElement('script');
-    script.src = browser.runtime.getURL('inject.js');
+    script.src = browser.runtime.getURL('script.js');
     const cmdString = commands.autoCommands.toString();
     script.onload = function () {
         const data = cmdString;
