+++ src/preload/extend_mathquill.js
