+++ src/script.ts
